node: Merge {insert,refresh}Neighbor()
Previously, refreshNeighbor() was very similar to insertNeighor().  This
commit refactors insertNeighbor() so that it can be invoked by
refreshNeighbor(), and thus DRY.

Besides refactoring, another notable changes: the linuxNodeHandler.mutex
is being held during the whole insertNeighor() execution. Previously,
refreshNeighbor() took the lock twice which opened a window for a race
when n.neighByNextHop[nextHopStr] could have been set to an obsolete
neigh if insertNeighor() and refreshNeighbor() were executed at the same
time.

Signed-off-by: Martynas Pumputis <m@lambda.lt>